chore: simplify footer copyright year output
Remove the unnecessary integer cast from the footer year rendering to keep the template cleaner while preserving the displayed copyright year.chore: simplify footer copyright year output

Remove the unnecessary integer cast from the footer year rendering to keep the template cleaner while preserving the displayed copyright year.

+++ src/app/views/layout/footer.php
@@ -34,7 +34,7 @@
     </nav>
 
     <div class="footer-bottom">
-      <p>&copy; <?= (int) date('Y') ?> Cartly. All rights reserved.</p>
+      <p>&copy; <?= date('Y') ?> Cartly. All rights reserved.</p>
       <p>Built for recipe planning, local shopping, and smarter carts.</p>
     </div>
   </div>
